POLIO-1187 refactor LQASAFroMApPanesContainer
- move getRounds to utils
- use context to get bounds and setBounds
